+++ app/__init__.py
@@ -7,7 +7,9 @@ from config import Config
 app = Flask(__name__)
 
 app.config.from_object(Config)
+
 bootstrap = Bootstrap(app)
+
 login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = 'login'
@@ -15,5 +17,4 @@ login_manager.login_view = 'login'
 db = SQLAlchemy(app)
 
 
-from app import old_app
 from app import routes
