+++ app/forms.py
@@ -1,0 +1,34 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, BooleanField
+from wtforms.validators import InputRequired, Email, Length, ValidationError
+
+from app.models import User
+
+
+class LoginForm(FlaskForm):
+    username = StringField('username', validators=[InputRequired()])
+    password = PasswordField('password', validators=[InputRequired()])
+    remember = BooleanField('remember me')
+
+
+class RegisterForm(FlaskForm):
+    email = StringField('email', validators=[
+        InputRequired(),
+        Email(message='Invalid email'),
+        Length(max=50)])
+    username = StringField('username', validators=[
+        InputRequired(),
+        Length(min=4, max=15)])
+    password = PasswordField('password', validators=[
+        InputRequired(),
+        Length(min=8, max=80)])
+
+    def validate_email(self, email):
+        users = User.query.filter_by(email=email.data).all()
+        if users:
+            raise ValidationError('This email already exists.')
+
+    def validate_username(self, username):
+        users = User.query.filter_by(username=username.data).all()
+        if users:
+            raise ValidationError('This username already exists.')
